Update nft-rfc-008.md
Editing the Verification Process section and Example to more closely align with the diagram shown by Joe on the 1/28 call. Including generalizing the Verification Process so it doesn't read as a slightly less specific example (should generalize better). 

TODO: as part of this edit the need for having explicit "VC1/2/3/4/5" seemed to fall out. They are still referenced in the example but most have been replaced by a name that can likely be used but this does have the result of maybe underspecifying that VCs are being used here. Need to revisit this shift as we do want the data models in use to be clear through the example.  Also the shift to capabilities from credential issuing has maybe eliminated the need for so many VC#s to appear.